Move routing to index
